server/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
tensorflow>=2.12
numpy
Pillow
python-multipart
